+++ crates/staged-sync/src/utils/mod.rs
@@ -2,4 +2,5 @@
 
 /// Utilities for initializing parts of the chain
 pub mod init;
+
 pub(crate) mod versioning;
